chore(wallet-core): utils cleanup
